+++ app/models/concerns/protip_spam.rb
@@ -1,0 +1,10 @@
+module ProtipSpam
+  extend ActiveSupport::Concern
+  included do
+    after_create :analyze_spam
+
+    def analyze_spam
+      AnalyzeSpamJob.perform_async({ id: id, klass: self.class.name })
+    end
+  end
+end
